+++ src/main/java/com/ols/ols_project/model/AccepteEntity.java
@@ -20,17 +20,17 @@ public class AccepteEntity {
     /**
      * 接受任务编号
      */
-    private Integer id;
+    private Long id;
 
     /**
      * 接受者编号
      */
-    private Integer user_id;
+    private Long user_id;
 
     /**
      * 任务编号
      */
-    private Integer task_id;
+    private Long task_id;
 
     /**
      * 接受时间
